Fixes to the trickle ice library and other misc changes
This commit includes fixes and changes to the trickle ice code and some
trivial changes in the platform library code. With these changes, we
were able to achieve the forwarding of broadcasted media to the receiver
and receiver chrome browser client was able to render audio and video.

One thing to remember w.r.t to trickle ice is that, so far we have
tested only connectivity between host candidates (everybody on same
lan). A lot of testing is required further for other nat scenarios.